+++ action_server/src/robocorp/action_server/_selftest.py
@@ -77,6 +77,10 @@ class ActionServerProcess:
         actions_sync=False,
         cwd: Optional[Path | str] = None,
         add_shutdown_api: bool = False,
+        min_processes: int = 0,
+        max_processes: int = 20,
+        reuse_processes: bool = False,
+        additional_args: Optional[list[str]] = None,
     ) -> None:
         from robocorp.action_server._robo_utils.process import Process
         from robocorp.action_server._settings import is_frozen
@@ -106,12 +110,21 @@ class ActionServerProcess:
             f"--datadir={str(self._datadir)}",
             f"--db-file={db_file}",
         ]
+
+        new_args.append(f"--min-processes={min_processes}")
+        new_args.append(f"--max-processes={max_processes}")
+        if reuse_processes:
+            new_args.append("--reuse-processes")
+
+        if additional_args:
+            new_args = new_args + additional_args
+
         env = {}
         if add_shutdown_api:
             env["RC_ADD_SHUTDOWN_API"] = "1"
         process = self._process = Process(new_args, cwd=cwd, env=env)
 
-        compiled = re.compile(r"http://([\w.-]+):(\d+)")
+        compiled = re.compile(r"Action Server started at http://([\w.-]+):(\d+)")
         future: Future[Tuple[str, str]] = Future()
 
         def collect_port_from_stdout(line):
@@ -133,7 +146,7 @@ class ActionServerProcess:
         process.on_stderr.register(on_stderr)
         process.on_stdout.register(on_stdout)
 
-        with process.on_stdout.register(collect_port_from_stdout):
+        with process.on_stderr.register(collect_port_from_stdout):
             process.start()
             if timeout > 1:
                 initial_time = time.monotonic()
@@ -194,10 +207,10 @@ class ActionServerClient:
         except Exception:
             raise AssertionError(f"Unable to load: {contents!r}")
 
-    def post_get_str(self, url, data):
+    def post_get_str(self, url, data, headers: Optional[dict] | None = None):
         import requests
 
-        result = requests.post(self.build_full_url(url), json=data)
+        result = requests.post(self.build_full_url(url), headers=headers, json=data)
         assert result.status_code == 200
         return result.text
 
@@ -206,6 +219,7 @@ class ActionServerClient:
 
         result = requests.post(self.build_full_url(url), json=data or {})
         assert result.status_code == status_code
+        return result
 
     def get_error(self, url, status_code):
         import requests